Increase number of rays cast, add keyboard shortcuts for bbh depth

+++ include/tracing/Benchmark.hpp
@@ -4,8 +4,8 @@ namespace benchmark
 {
     struct HitTests
     {
-        int triangle_tests = 0;
-        int bbox_tests = 0;
+        int64_t triangle_tests = 0;
+        int64_t bbox_tests = 0;
 
         constexpr void registerTriangleTest() { ++triangle_tests; }
         constexpr void registerBBoxTest() { ++bbox_tests; }
